EnergyBarWithCharacter

+++ Assets/Junseo/UI Script/PlayerStatusInterface.cs
@@ -1,5 +1,5 @@
 public interface PlayerStatusInterface
 {
-    int GetEnergy();             // 에너지 값 가져오기
-    void UseEnergy(int amount); // 에너지 사용
+    float GetEnergy();             // 에너지 값 가져오기
+    void UseEnergy(float amount); // 에너지 사용
 }
